add LargeBotQueue as supported queue backend (#2)
Add `LargeBotQueue` as a non-default supported backend. This alternative
backend will be used if a `DISCORD_TOKEN` enviroment variable is
provided, and if not, a `LocalQueue` will be used. It will fetch
`max_concurrency` from the Discord API and build the appropiate queue.

This is backwards compatible as the old behavior is the same as the new
one as this will only be used if the enviroment variable is present.

```
$ PORT=8080 ./target/release/twilight-gateway-queue
 2020-10-16T11:42:35.215Z INFO  twilight_gateway_queue > Listening on http://0.0.0.0:8080
 2020-10-16T11:42:42.404Z INFO  twilight_gateway_queue > shard 0/0 waiting for allowance
$ PORT=8080 DISCORD_TOKEN="snip" ./target/release/twilight-gateway-queue
 2020-10-16T11:42:56.412Z INFO  twilight_gateway_queue > Listening on http://0.0.0.0:8080
 2020-10-16T11:42:58.206Z INFO  twilight_gateway_queue::large_bot_queue > waiting for allowance on shard 0
```

Approved-by: Valdemar Erk <valdemar@erk.io>
Approved-by: Vivian Hellyer <vivian@hellyer.dev>
Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Jens Reidel <adrian@travitia.xyz>